Fix CI, simplify sidebar
