--- frontend/src/common/sidebar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './sidebar';
